Refactor: Refactor general code throughout mod.

+++ p5rpc.modloader/Patches/P5R/SkipIntro.cs
@@ -15,7 +15,7 @@ internal class SkipIntro
             return;
         
         context.ScanHelper.FindPatternOffset("74 10 C7 07 0C 00 00 00", (offset) => 
-            Memory.Instance.SafeWriteRaw((nuint)((nint)baseAddr + offset), new byte[] { 0x90, 0x90 }),
+            Memory.Instance.SafeWriteRaw((nuint)(baseAddr + offset), new byte[] { 0x90, 0x90 }),
             "Introduction Skip");
     }
 }
